x1_012: Fix insidious variable type mistake that caused tilemap regressions

+++ src/mame/video/x1_012.h
@@ -48,7 +48,7 @@ private:
 
 	tilemap_t *m_tilemap;
 
-	u8 m_xoffsets[2];
+	int m_xoffsets[2];
 
 	u16 m_vctrl[3];
 	u8 m_rambank;
